requirements.txt
streamlit
pandas
plotly
pygwalker
openai
opencv-python-headless
tavily-python
asyncio
bs4
requests    
supabase
urllib3
streamlit_elements
streamlit_extras
extra_streamlit_components
temp
opencv-python-headless
ipython
pillow
numpy
pytube
pydub
youtube-transcript-api
moviepy